new function set working (tests pass)
- bugfixes in interpreter for relational operators
- project file tidied up
- small memory allocation bugfixes
- bugfix with processing qrng location

+++ EACirc/generators/QuantumRndGen.cpp
@@ -30,7 +30,7 @@ QuantumRndGen::~QuantumRndGen() {
 
 void QuantumRndGen::checkQRNGdataAvailability() {
     int length;
-    unsigned int separatorPosition;
+    size_t separatorPosition;
     string paths = m_QRNGDataPath;
     while (!paths.empty()) {
         string currentPath = paths.substr(0,paths.find(FILE_PATH_SEPARATOR));
